requirements.txt
pushbullet.py==0.12.0
python-dotenv==0.19.2